requirements.txt
Flask==2.2.5
gunicorn==20.1.0
PyMySQL==1.1.0
requests==2.31.0
python-dotenv==1.0.0
Jinja2==3.1.2
Werkzeug==2.2.3 